2023 day 1
Part 1 was easy as you could just match the first and last digits and
they have no overlap.

Part 2 took me longer than it should because I tried to replace the
words with numbers and then get the first and last ones.
This really broke with overlapping numbers like `twone` specifically when finding the last
number. Also annoyingly the it could correctly generate the 281 from the
sample test data. To fix this I changed tack and searched for numbers or
words at the same time and then replaced the words during the addition.